chore: add default map
