+++ vaadin-map-flow-parent/vaadin-map-flow/src/main/java/com/vaadin/flow/component/map/Map.java
@@ -1,94 +1,50 @@
 package com.vaadin.flow.component.map;
 
-import com.vaadin.flow.component.AttachEvent;
-import com.vaadin.flow.component.Component;
-import com.vaadin.flow.component.HasSize;
 import com.vaadin.flow.component.Tag;
 import com.vaadin.flow.component.dependency.JsModule;
 import com.vaadin.flow.component.map.configuration.Configuration;
-import com.vaadin.flow.component.map.configuration.View;
-import com.vaadin.flow.internal.JsonSerializer;
-import com.vaadin.flow.internal.StateTree;
-import elemental.json.JsonObject;
-
-import java.beans.PropertyChangeEvent;
+import com.vaadin.flow.component.map.configuration.layer.Layer;
+import com.vaadin.flow.component.map.configuration.layer.TileLayer;
+import com.vaadin.flow.component.map.configuration.source.OSMSource;
 
 @Tag("vaadin-map")
 // TODO: Enable once released
 // @NpmPackage(value = "@vaadin/map", version = "23.0.0-alpha4")
 // TODO: Include non-themed module `@vaadin/map/src/vaadin-map.js` when theme module is ready
 @JsModule("@vaadin/map/vaadin-map.js")
-public class Map extends Component implements HasSize {
+public class Map extends MapBase {
 
-    private Configuration configuration;
-    private View view;
-
-    private StateTree.ExecutionRegistration pendingConfigurationSync;
-    private StateTree.ExecutionRegistration pendingViewSync;
+    private Layer baseLayer;
 
     public Map() {
-        this.configuration = new Configuration();
-        this.view = new View();
-        this.configuration.addPropertyChangeListener(this::configurationPropertyChange);
-        this.view.addPropertyChangeListener(this::viewPropertyChange);
+        super();
+        OSMSource source = new OSMSource();
+        TileLayer baseLayer = new TileLayer();
+        baseLayer.setSource(source);
+        setBaseLayer(baseLayer);
     }
 
-    public Configuration getConfiguration() {
-        return configuration;
+    public Configuration getRawConfiguration() {
+        return getConfiguration();
     }
 
-    public View getView() {
-        return view;
+    public Layer getBaseLayer() {
+        return baseLayer;
     }
 
-    public void render() {
-        this.requestConfigurationSync();
-    }
-
-    @Override
-    protected void onAttach(AttachEvent attachEvent) {
-        super.onAttach(attachEvent);
-        requestConfigurationSync();
-        requestViewSync();
-    }
-
-    private void requestConfigurationSync() {
-        if (pendingConfigurationSync != null) {
-            return;
+    public void setBaseLayer(Layer baseLayer) {
+        if (this.baseLayer != null) {
+            getConfiguration().removeLayer(this.baseLayer);
         }
-        getUI().ifPresent(ui -> pendingConfigurationSync = ui.beforeClientResponse(this, context -> {
-            pendingConfigurationSync = null;
-            synchronizeConfiguration();
-        }));
+        this.baseLayer = baseLayer;
+        getConfiguration().prependLayer(baseLayer);
     }
 
-    private void requestViewSync() {
-        if (pendingViewSync != null) {
-            return;
-        }
-        getUI().ifPresent(ui -> pendingViewSync = ui.beforeClientResponse(this, context -> {
-            pendingViewSync = null;
-            synchronizeView();
-        }));
+    public void addLayer(Layer layer) {
+        getConfiguration().addLayer(layer);
     }
 
-    private void synchronizeConfiguration() {
-        JsonObject jsonConfiguration = (JsonObject) JsonSerializer.toJson(configuration);
-
-        this.getElement().executeJs("this.updateConfigurationJson($0)", jsonConfiguration);
-    }
-
-    private void synchronizeView() {
-        JsonObject jsonView = (JsonObject) JsonSerializer.toJson(view);
-
-        this.getElement().executeJs("this.updateViewJson($0)", jsonView);
-    }
-
-    private void configurationPropertyChange(PropertyChangeEvent e) {
-        this.requestConfigurationSync();
-    }
-
-    private void viewPropertyChange(PropertyChangeEvent e) {
-        this.requestViewSync();
+    public void removeLayer(Layer layer) {
+        getConfiguration().removeLayer(layer);
     }
 }
